+++ src/request.ts
@@ -43,6 +43,11 @@ export interface RevocationRequest {
   data: RevocationRequestData;
 }
 
+export interface OffchainRevocationRequest {
+  schema: string;
+  data: RevocationRequestData;
+}
+
 export interface DelegatedRevocationRequest extends RevocationRequest {
   signature: Signature;
   revoker: string;
@@ -60,7 +65,7 @@ export interface MultiDelegatedRevocationRequest extends MultiRevocationRequest 
   deadline?: bigint;
 }
 
-export interface DelegatedProxyAttestationRequest extends DelegatedAttestationRequest {}
-export interface MultiDelegatedProxyAttestationRequest extends MultiDelegatedAttestationRequest {}
-export interface DelegatedProxyRevocationRequest extends DelegatedRevocationRequest {}
-export interface MultiDelegatedProxyRevocationRequest extends MultiDelegatedRevocationRequest {}
+export type DelegatedProxyAttestationRequest = DelegatedAttestationRequest;
+export type MultiDelegatedProxyAttestationRequest = MultiDelegatedAttestationRequest;
+export type DelegatedProxyRevocationRequest = DelegatedRevocationRequest;
+export type MultiDelegatedProxyRevocationRequest = MultiDelegatedRevocationRequest;