+++ src/schema-registry.ts
@@ -2,10 +2,10 @@ import {
   SchemaRegistry__factory,
   SchemaRegistry as SchemaRegistryContract
 } from '@ethereum-attestation-service/eas-contracts';
-import { Overrides, TransactionReceipt } from 'ethers';
+import { Overrides, solidityPackedKeccak256, TransactionReceipt } from 'ethers';
 import { legacyVersion } from './legacy/version';
-import { Base, Transaction, TransactionSigner } from './transaction';
-import { getSchemaUID, ZERO_ADDRESS, ZERO_BYTES32 } from './utils';
+import { Base, RequireSigner, Transaction, TransactionProvider, TransactionSigner } from './transaction';
+import { ZERO_ADDRESS, ZERO_BYTES32 } from './utils';
 
 export declare type SchemaRecord = {
   uid: string;
@@ -25,7 +25,7 @@ export interface GetSchemaParams {
 }
 
 export interface SchemaRegistryOptions {
-  signer?: TransactionSigner;
+  signer?: TransactionSigner | TransactionProvider;
 }
 
 export class SchemaRegistry extends Base<SchemaRegistryContract> {
@@ -40,20 +40,22 @@ export class SchemaRegistry extends Base<SchemaRegistryContract> {
     return (await legacyVersion(this.contract)) ?? this.contract.version();
   }
 
+  // Returns a schema UID
+  public static getSchemaUID(schema: string, resolverAddress: string, revocable: boolean) {
+    return solidityPackedKeccak256(['string', 'address', 'bool'], [schema, resolverAddress, revocable]);
+  }
+
   // Registers a new schema and returns its UID
+  @RequireSigner
   public async register(
     { schema, resolverAddress = ZERO_ADDRESS, revocable = true }: RegisterSchemaParams,
     overrides?: Overrides
   ): Promise<Transaction<string>> {
-    if (!this.signer) {
-      throw new Error('Invalid signer');
-    }
-
     return new Transaction(
       await this.contract.register.populateTransaction(schema, resolverAddress, revocable, overrides ?? {}),
-      this.signer,
+      this.signer!,
       // eslint-disable-next-line require-await
-      async (_receipt: TransactionReceipt) => getSchemaUID(schema, resolverAddress, revocable)
+      async (_receipt: TransactionReceipt) => SchemaRegistry.getSchemaUID(schema, resolverAddress, revocable)
     );
   }